fix: close 'register dApp modal'
