+++ src/components/dapp-staking/modals/StakeModal.vue
@@ -40,9 +40,10 @@
         {{ $t('dappStaking.yourStake') }}
         <format-balance :balance="stakeAmount" class="tw-inline tw-font-semibold" />
       </div>
-    </template>
-    <template #buttons>
-      <Button :disabled="!canExecuteAction" @click="action(data)">{{ actionName }}</Button>
+      <div class="tw-mt-6 tw-flex tw-justify-center tw-flex-row">
+        <Button type="button" :primary="false" @click="closeModal">{{ $t('close') }}</Button>
+        <Button :disabled="!canExecuteAction" @click="action(data)">{{ actionName }}</Button>
+      </div>
     </template>
   </Modal>
 </template>
@@ -101,7 +102,8 @@ export default defineComponent({
       required: true,
     },
   },
-  setup(props) {
+  emits: ['update:is-open'],
+  setup(props, { emit }) {
     const store = useStore();
     const { decimal, defaultUnitToken } = useChainMetadata();
 
@@ -140,6 +142,10 @@ export default defineComponent({
       store.commit('general/setCurrentAccountIdx', selAccountIdx);
     };
 
+    const closeModal = () => {
+      emit('update:is-open', false);
+    };
+
     return {
       data,
       allAccounts,
@@ -148,6 +154,7 @@ export default defineComponent({
       reloadAmount,
       StakeAction,
       canExecuteAction,
+      closeModal,
       ...toRefs(props),
     };
   },
